+++ packages/widgetbook_generator/lib/src/generators/addons_configs_builder.dart
@@ -66,14 +66,14 @@ class AddonsConfigsBuilder extends BaseBuilder {
 
             if (key == 'viewport') {
               final data = reader.read('value');
-              final id = data.read('id').stringValue;
+              final name = data.read('name').stringValue;
               final ratio = data.read('pixelRatio').doubleValue;
               final width = data.read('width').doubleValue;
               final height = data.read('height').doubleValue;
 
               return AddonConfig(
                 key,
-                'id:$id,\$meta:{w:$width,h:$height,p:$ratio}',
+                'name:$name,\$meta:{w:$width,h:$height,p:$ratio}',
               );
             }